refactor: colocate schema data models with their JSON schemas
Move finding models and metadata from root types.py to schemas/types.py
for proper collocation with their JSON schema definitions.

Changes:
- PersonalDataAnalyser: Move PersonalDataFindingModel and PersonalDataFindingMetadata to schemas/types.py
- DataSubjectAnalyser: Move DataSubjectFindingModel and DataSubjectFindingMetadata to schemas/types.py
- ProcessingPurposeAnalyser: Move ProcessingPurposeFindingModel and ProcessingPurposeFindingMetadata to schemas/types.py
- Update all internal imports to use .schemas.types for finding models
- Update test imports to use correct paths
- Update prompt imports to use .schemas package
- Maintain config classes in root types.py for proper separation

This achieves clean separation of concerns:
- Root types.py: Configuration models (*AnalyserConfig)
- schemas/types.py: Schema data models (*FindingModel, *FindingMetadata)

All 752 tests passing with full type checking coverage.
