+++ libs/waivern-community/src/waivern_community/analysers/data_subject_analyser/__init__.py
@@ -1,7 +1,8 @@
 """Data subject analyser for GDPR Article 30(1)(c) compliance."""
 
 from .analyser import DataSubjectAnalyser
-from .types import DataSubjectAnalyserConfig, DataSubjectFindingModel
+from .schemas import DataSubjectFindingModel
+from .types import DataSubjectAnalyserConfig
 
 __all__ = [
     "DataSubjectAnalyser",
