+++ libs/waivern-community/src/waivern_community/analysers/data_subject_analyser/schemas/__init__.py
@@ -3,7 +3,11 @@
 from waivern_community.analysers.data_subject_analyser.schemas.data_subject_finding import (
     DataSubjectFindingSchema,
 )
+from waivern_community.analysers.data_subject_analyser.schemas.types import (
+    DataSubjectFindingModel,
+)
 
 __all__ = [
+    "DataSubjectFindingModel",
     "DataSubjectFindingSchema",
 ]
