+++ libs/waivern-community/src/waivern_community/analysers/data_subject_analyser/types.py
@@ -1,13 +1,12 @@
-"""Types for data subject analyser."""
+"""Configuration types for data subject analyser."""
 
 from typing import Any, Self
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import BaseModel, Field
 from waivern_analysers_shared.types import (
     LLMValidationConfig,
     PatternMatchingConfig,
 )
-from waivern_core.schemas import BaseFindingModel
 
 
 class DataSubjectAnalyserConfig(BaseModel):
@@ -38,27 +37,3 @@ class DataSubjectAnalyserConfig(BaseModel):
 
         """
         return cls.model_validate(properties)
-
-
-class DataSubjectFindingMetadata(BaseModel):
-    """Metadata for data subject findings."""
-
-    source: str = Field(description="Source file or location where the data was found")
-
-    model_config = ConfigDict(extra="forbid")
-
-
-class DataSubjectFindingModel(BaseFindingModel):
-    """Data subject finding structure."""
-
-    primary_category: str = Field(description="Primary data subject category")
-    confidence_score: int = Field(
-        ge=0, le=100, description="Confidence score for the classification (0-100)"
-    )
-    modifiers: list[str] = Field(
-        default_factory=list,
-        description="Cross-category regulatory modifiers from ruleset",
-    )
-    metadata: DataSubjectFindingMetadata | None = Field(
-        default=None, description="Additional metadata about the finding"
-    )
