+++ libs/waivern-community/src/waivern_community/analysers/processing_purpose_analyser/__init__.py
@@ -1,6 +1,6 @@
 """Processing purpose analyser for GDPR compliance analysis."""
 
 from .analyser import ProcessingPurposeAnalyser
-from .types import ProcessingPurposeFindingModel
+from .schemas import ProcessingPurposeFindingModel
 
 __all__ = ["ProcessingPurposeAnalyser", "ProcessingPurposeFindingModel"]
