+++ libs/waivern-community/src/waivern_community/analysers/processing_purpose_analyser/schemas/__init__.py
@@ -3,7 +3,11 @@
 from waivern_community.analysers.processing_purpose_analyser.schemas.processing_purpose_finding import (
     ProcessingPurposeFindingSchema,
 )
+from waivern_community.analysers.processing_purpose_analyser.schemas.types import (
+    ProcessingPurposeFindingModel,
+)
 
 __all__ = [
+    "ProcessingPurposeFindingModel",
     "ProcessingPurposeFindingSchema",
 ]
