+++ libs/waivern-community/src/waivern_community/analysers/processing_purpose_analyser/source_code_schema_input_handler.py
@@ -12,7 +12,10 @@ from waivern_community.connectors.source_code.schemas import (
     SourceCodeFileDataModel,
 )
 
-from .types import ProcessingPurposeFindingMetadata, ProcessingPurposeFindingModel
+from .schemas.types import (
+    ProcessingPurposeFindingMetadata,
+    ProcessingPurposeFindingModel,
+)
 
 
 class SourceCodeFileMetadata(BaseModel):
