+++ libs/waivern-community/tests/waivern_community/analysers/processing_purpose_analyser/test_pattern_matcher.py
@@ -11,7 +11,7 @@ from waivern_analysers_shared.types import PatternMatchingConfig
 from waivern_community.analysers.processing_purpose_analyser.pattern_matcher import (
     ProcessingPurposePatternMatcher,
 )
-from waivern_community.analysers.processing_purpose_analyser.types import (
+from waivern_community.analysers.processing_purpose_analyser.schemas.types import (
     ProcessingPurposeFindingMetadata,
     ProcessingPurposeFindingModel,
 )
