+++ libs/waivern-personal-data-analyser/src/waivern_personal_data_analyser/schemas/__init__.py
@@ -1,7 +1,7 @@
 """Schema for personal data analyser output."""
 
-from ..types import PersonalDataFindingModel
 from .personal_data_finding import PersonalDataFindingSchema
+from .types import PersonalDataFindingModel
 
 __all__ = [
     "PersonalDataFindingModel",
